consolidate plugins further
