+++ fastifyPlugins/main.js
@@ -1,4 +1,45 @@
+// Hypns will connect us to the network and pin the data
+const HyPNS = require('hypns')
+// Security note: It's deliberately placed in the `.data` directory which doesn't get copied if someone remixes the project.
+const hypnsNode = new HyPNS({ persist: true, applicationName: '.data/hypnsapp' })
+
 module.exports = function (fastify, options, done) {
+  const setUp = async (publicKey) => {
+    console.log('setting up ', publicKey)
+
+    const instance = await hypnsNode.open({ keypair: { publicKey } })
+    await instance.ready()
+    console.log(`1. Listener count ${instance.listenerCount('update')} on ${instance.key}`)
+
+    // skip if the instance is already listening
+    if (instance.listenerCount('update') < 1) {
+      console.log('Setup ', instance.publicKey, ` latest:${instance.listenerCount('update')}`)
+      const d = Date.now()
+      instance.on('update', (val) => {
+        const lag = (new Date(Date.now())) - (new Date(instance.latest.timestamp))
+        console.log('Update ', instance.publicKey, ` latest:${instance.latest.timestamp} ${instance.latest.text} [${new Date(lag).getSeconds()} sec] on ${d}`)
+        fastify.db.set(`pins.${publicKey}`, instance.latest)
+          .write()
+      })
+    }
+    console.log(`2. Listener count ${instance.listenerCount('update')} on ${instance.key}`)
+
+    fastify.db.set(`pins.${publicKey}`, instance.latest).write()
+    console.log('** Setup COMPLETE ** \n', instance.publicKey, ` pins.size: [${fastify.db.get('pins').size().value()}]`)
+    return instance.latest
+  }
+
+  // load list from storage and initialize the node
+  const init = async () => {
+    const pins = fastify.db.get('pins').value() // Find all publicKeys pinned in the collection
+    Object.keys(pins).forEach((key) => {
+      // skip if it's already configured on this node
+      if (!hypnsNode.instances.has(key)) { setUp(key) }
+    })
+  }
+
+  init()
+
   fastify.register((fi, options, done) => {
     // https://www.fastify.io/docs/latest/Validation-and-Serialization/
     const opts = {
@@ -38,7 +79,7 @@ module.exports = function (fastify, options, done) {
   fastify.get('/pins/',
     async (request, reply) => {
       let out = ''
-      for (const inst of fastify.instances.values()) {
+      for (const inst of hypnsNode.instances.values()) {
         if (inst.latest) {
           out += `\n<br />${inst.latest.timestamp} ${inst.publicKey}: ${inst.latest.text}`
         } else {
@@ -58,7 +99,7 @@ module.exports = function (fastify, options, done) {
   fastify.get('/latest/', { schema: { querystring: { rootKey: { type: 'string' } } } },
     async function (request, reply) {
       const publicKey = request.querystring.rootKey
-      const latest = fastify.getLatest(publicKey)
+      const latest = hypnsNode.instances.get(publicKey).latest
       console.log(`** GET Latest ${publicKey}: ${latest}`)
       reply.send({ latest })
     }
