+++ fastifyPlugins/static.js
@@ -10,5 +10,22 @@ module.exports = function (fastify, options, done) {
     reply.sendFile('index.html')
   })
 
+  fastify.get('/pins', function (request, reply) {
+    console.log('*** getting pins ***')
+    const pins = fastify.db.get('pins').value() // Find all publicKeys pinned in the collection
+    console.log('*** pins ***', pins)
+    reply.send(pins) // sends pins back to the page
+  })
+
+  // removes all entries from the collection
+  fastify.get('/clear', function (request, reply) {
+    // removes all entries from the collection
+    fastify.db.get('pins')
+      .remove()
+      .write()
+    console.log('Database cleared')
+    reply.redirect('/')
+  })
+
   done()
 }
